# pkg/idutil: use count field as atomic variable