Use atomic.AddUint64(_fetch-and-add primitive_) instead of mutex lock to inc count field.

Bench result:
```
benchmark           old ns/op     new ns/op     delta
BenchmarkNext-4     163           26.3          -83.87%
```