@lorneli 2018-04-26
Maybe it's worthy to optimize because generating unique id is high frequency operation.